+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Declare the originating types for generated `*Builder` class to make it easier for tooling to determine when to re-run annotation processor to generate builder class.
+* Arez removed support for using non-constructor based injection of services in version `0.153` propagated the same restrictions to React4j. The annotation processor code to support field and method injections within React4j has now been removed to bring it inline
 
 ### [v0.139](https://github.com/react4j/react4j/tree/v0.139) (2019-11-05) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.138...v0.139)